Ensure patron logged out of CAS after vega logout
The Vega Logout route does not always route the user through the CAS
Logout endpoint. This means that, after a user passes through the Vega
Logout route, they may still be logged in to CAS. One way to ensure they
are not logged in to CAS is to pass them through the CAS Logout route
after they've gone through the Vega Logout route.

Adds a new route to the RedirectService:
/vega-logout-handler[?redirect_uri=...] . The new route accepts a valid
redirect_uri and passes the user on through the CAS Logout route.

Modifies the Encore Logout handler to, before passing the patron on to the
Vega Logout route, modifies the incoming redirect_uri to first hit the new
Vega logout handler route so that the patron routes through CAS Logout
after logging out of Vega.

This should ensure that patrons are logged out of both CAS and Vega
after logging out of RC.