Remove useless, confusing if statement
Remove a confusing if statement that is missing a block, causing it to
join the subsequent `if`, creating a compound clause that will never
evaluate as `true` (because the request domain will never be the Vega
domain since this app doesn't serve that host). As a consequence, the
lines appear to be useless.